Enhanced Kodi app
This application is a little more resource-heavy:
Kodi's JSON-RPC does not have a sort of "stats" call, so we need to call individual library components to get the total numbers. This can become network-heavy if you have a gigantic library, I guess.

To-Do: option to replace the library stats with "Now playing" info, in case there's an active player.