Make missing stat more meaningful
Current missing stat does not reflect any built in filters of the movie list in Radarr.
Monitored, available and no files is the same as the "Wanted" view. This is equivalent to the missing episodes is Sonarr.

Radarr's "Missing" view would be `$missing = $collect->where('monitored', true)->where('hasFile', false);` . But I prefer Wanted as those are the movies you could actually get.