fix: Fix *arr url creation (#788)
* Update Radarr.php

Radarr was using normaliseurl (which strips the trailing `/`) without passing true to re-add said `/` as needed.

* Update Sonarr.php

Sonarr as well was using normaliseurl (which strips the trailing `/`) without passing true to re-add said `/` as needed.

* Update Readarr.php

Readarr  as well was using normaliseurl (which strips the trailing `/`) without passing true to re-add said `/` as needed.

* Update Lidarr.php

Lidarr as well was using normaliseurl (which strips the trailing `/`) without passing true to re-add said `/` as needed.

* Update Bazarr.php

Bazarr as well was using normaliseurl (which strips the trailing `/`) without passing true to re-add said `/` as needed.

* Update Prowlarr.php

Prowlarr as well was using normaliseurl (which strips the trailing `/`) without passing true to re-add said `/` as needed.