Immich: display large numbers with suffixes
Because an app plate is quite small, displaying large numbers
can be problematic as they collide and cannot be distinguished.

This commit fixes this issue by displaying large numbers using
fewer characters and ISO suffixes, up to G (giga, 10^9).

Here is an example of the problematic behaviour:
+-----------------------+
| PHOTOS  VIDEOS  USAGE |
|   14398443 0.94GiB    |
+-----------------------+
The 14,398 photo count collides with the 443 video count.

This commit changes the display to:
+-----------------------+
| PHOTOS  VIDEOS  USAGE |
|   15k    443  0.94GiB |
+-----------------------+