Immich: Fix blade's coliding titles
Immich blades displays three metrics: one for the number of photos,
one for the number of videos and one for the total space usage of
its server.

This commit fixes an issue where the titles of those metrics are
colliding into a single word: "PHOTOSVIDEOsUSAGE", which impair
readability.

Before this commit, the blade reads:
<pre>
+-------------------+
| PHOTOSVIDEOSUSAGE |
|   265 443 0.94GiB |
+-------------------+
</pre>

After this commit, the blade reads:
<pre>
+-----------------------+
| PICS    VIDS    USAGE |
|  265    443   0.94GiB |
+-----------------------+
</pre>